求解超时设置调整

+++ natives/optimal/src/main/kotlin/top/bettercode/summer/tools/optimal/solver/COPTSolver.kt
@@ -33,7 +33,7 @@ class COPTSolver @JvmOverloads constructor(
         model.setDblParam(copt.DblParam.FeasTol, 1e-9)
     }
 
-    override fun setTimeLimit(seconds: Int) {
+    override fun setTimeLimit(seconds: Long) {
         model.setDblParam("TimeLimit", seconds.toDouble())
     }
